softwareUpdate.sh
The goal of this script is to leverage Jamf Helper notification windows to present end users with an opt-in and opt-out option to perform the available software updates. Before the end user is presented with the Jamf Helper notification, a series of checks are performed to ensure if updates are still available. This will also parse through which require restarts and which do not. Depending on this, a restart will be conducted should the user elect to go forward with the updates.